refactor(products): decompose ProductsStockService into Transfer/Overview/Reservation services
Behavior-preserving (write-heavy, jeRefs=0, NO existing spec → strictly verbatim).
Facade keeps 15-method surface + 1-arg ctor + delegates; internally constructs 3
sub-services (module untouched). 5 $transaction (incl bulkTransfer 15s timeout +
confirmTransfer atomic branchId move) moved WHOLE + byte-identical. tsc 0, lint 0.
